ZSTDMT : implemented "getNewBuffer after previous job is accepted"
This is a bit complex to explain.

The way the ZSTDMT streamer creates a new job is :
- get a buffer
- fill it with input
- once appropriately filled :
  - get a new buffer for next job
  - fill its beginning with end of previous buffer
  - post a job with old buffer
  - block there until a worker thread is ready to accept the job
- continue with new buffer
It works, but during the blocking period, we have 2 buffers reserved.

The new method is as follows :
- once buffer is appropriately filled :
  - post a job with current buffer
  - block, until a worker is available
  - on return, request a new buffer
    - and fill its beginning with end of old buffer

In theory, it's better because during the wait period, only 1 buffer was reserved.
The expectation is that, when a worker is done, it releases its own buffer.
So when the main thread asks for a new buffer to prepare next job, it gets the newly release one.
This should decrease the total amount of buffers allocated by one.
Savings are 2 MB at level 1, 4 MB at level 3, 8 MB at level 6, etc. so not negligible.

Unfortunately, in practice, this new code is not so clear cut.
The new version "tends to" use one less buffer than previous version,
but it's just a mild tendency, not a clear win.
(tested with `fuzzer --memtest`)

There is also the problem of added complexity :
in order to read from old buffer, the main thread must ensure this buffer is still valid.
The worker thread itself only uses it for reading, so there is no race condition risk.
But when it's done, the buffer is released to the pool.
In general, it means the buffer is just transferred to the pool, while waiting for a new demand.
But if the pool is full, the buffer is freed instead.
In which case, the main thread would try to access data from a forbidden memory area.

To avoid that, a mutex is added into each job.
The main thread only release the mutex once it has finished copying the end of old buffer.
The worker thread can only release the buffer after acquiring the mutex.
So it all works well.

But now, it's also more complex.
There is one mutex to init/destroy per job, so the job table is no longer a clean data table,
it needs an init, hence a constructor, and a destructor.
So everything is a bit more complex now.
All this for a "mild" gain...

If that's all there is, it's probably not worth it.
But I'm still unsure if I properly implemented the idea, or if the execution was flawed...